q3/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
e2b==0.16.0
chromadb==0.4.15
groq==0.4.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.3
langchain-chroma==0.1.0
langgraph==0.0.20
langchain-text-splitters==0.0.1
unstructured[all-docs]==0.11.8
tiktoken==0.5.2 